suite refactor
